+++ src/client.ts
@@ -63,7 +63,6 @@ export class PirschNodeApiClient extends PirschCoreClient {
         const element: PirschHit = {
             url: url.toString(),
             ip: request.socket.remoteAddress ?? "",
-            dnt: this.getHeader(request.headers, "dnt"),
             user_agent: this.getHeader(request.headers, "user-agent") ?? "",
             accept_language: this.getHeader(request.headers, "accept-language"),
             sec_ch_ua: this.getHeader(request.headers, "Sec-CH-UA"),